requirements.txt
amqp==5.0.2
appnope==0.1.0
backcall==0.2.0
billiard==3.6.3.0
celery==5.0.2
click==7.1.2
click-didyoumean==0.0.3
click-repl==0.1.6
decorator==4.4.2
Django==2.2
django-celery-beat==2.1.0
django-celery-results==2.0.0
django-rest-framework==0.1.0
django-timezone-field==4.0
djangorestframework==3.12.2
importlib-metadata==3.1.0
ipdb==0.13.4
ipython==7.19.0
ipython-genutils==0.2.0
jedi==0.17.2
kombu==5.0.2
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.8
psycopg2==2.8.6
ptyprocess==0.6.0
Pygments==2.7.2
python-crontab==2.5.1
python-dateutil==2.8.1
pytz==2020.4
redis==3.5.3
six==1.15.0
sqlparse==0.4.1
traitlets==5.0.5
vine==5.0.0
wcwidth==0.2.5
zipp==3.4.0
